refactor: move getPoem logic to own file
